requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic 
passlib[bcrypt] 
pyjwt